[feat] Refactor AppProvider component and update imports
- Renamed the `AppProvider` component to `index` to streamline the codebase.
- Updated the import statement in root level `index.tsx` to reflect the rename of the component file.
- Changed the type of `children` in `AppProviderProps` from `React.ReactChild | React.ReactFragment` to `React.ReactNode` for better type flexibility.
